requirements.txt
yfinance
requests
beautifulsoup4
openai
python-dotenv
gradio